Lookup component without using css
